Break up lexing and layout stages
