+++ parser/src/grammar.lalrpop
@@ -3,8 +3,9 @@ use base::kind::Kind;
 use base::pos::{self, BytePos, Spanned};
 use base::types::{Alias, ArcType, BuiltinType, Field, Generic, Type};
 use combine::primitives::Error;
-use lexer::{Delimiter, Token};
 use std::str::FromStr;
+
+use token::{Delimiter, Token};
 
 use {FieldExpr, FieldPattern, MutIdentEnv, ErrorEnv};
